# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Beneficios - Kuantaz",
        "description": "API para gestionar beneficios procesados con filtros y fichas. Esta API consume datos de fuentes externas, aplica filtros de montos mínimos y máximos, y agrupa los beneficios por año con sus respectivas fichas asociadas.",
        "contact": {
            "name": "César Bugueno",
            "email": "buguenocesar92@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de Desarrollo Local"
        }
    ],
    "paths": {
        "/api/v1/beneficios-procesados": {
            "get": {
                "tags": [
                    "Beneficios"
                ],
                "summary": "Obtener beneficios procesados agrupados por año",
                "description": "Retorna los beneficios agrupados por año con filtros aplicados por montos mínimos y máximos. Cada beneficio incluye su ficha asociada con información detallada del programa.",
                "operationId": "getBeneficiosProcesados",
                "responses": {
                    "200": {
                        "description": "Lista de beneficios procesados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "ano": {
                                                "description": "Año de los beneficios",
                                                "type": "string",
                                                "example": "2023"
                                            },
                                            "total_monto": {
                                                "description": "Suma total de montos para el año",
                                                "type": "integer",
                                                "example": 295608
                                            },
                                            "num": {
                                                "description": "Número total de beneficios para el año",
                                                "type": "integer",
                                                "example": 8
                                            },
                                            "beneficios": {
                                                "description": "Lista de beneficios del año ordenados por fecha",
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id_programa": {
                                                            "description": "ID del programa de beneficio",
                                                            "type": "integer",
                                                            "example": 147
                                                        },
                                                        "monto": {
                                                            "description": "Monto del beneficio",
                                                            "type": "integer",
                                                            "example": 40656
                                                        },
                                                        "fecha_recepcion": {
                                                            "description": "Fecha de recepción del beneficio",
                                                            "type": "string",
                                                            "example": "09/11/2023"
                                                        },
                                                        "fecha": {
                                                            "description": "Fecha del beneficio en formato ISO",
                                                            "type": "string",
                                                            "example": "2023-11-09"
                                                        },
                                                        "ano": {
                                                            "description": "Año extraído de la fecha",
                                                            "type": "string",
                                                            "example": "2023"
                                                        },
                                                        "view": {
                                                            "description": "Indicador de visibilidad",
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "ficha": {
                                                            "description": "Información detallada del programa",
                                                            "properties": {
                                                                "id": {
                                                                    "description": "ID único de la ficha",
                                                                    "type": "integer",
                                                                    "example": 922
                                                                },
                                                                "nombre": {
                                                                    "description": "Nombre del programa",
                                                                    "type": "string",
                                                                    "example": "Emprende"
                                                                },
                                                                "id_programa": {
                                                                    "description": "ID del programa asociado",
                                                                    "type": "integer",
                                                                    "example": 147
                                                                },
                                                                "url": {
                                                                    "description": "URL slug del programa",
                                                                    "type": "string",
                                                                    "example": "emprende"
                                                                },
                                                                "categoria": {
                                                                    "description": "Categoría del programa",
                                                                    "type": "string",
                                                                    "example": "trabajo"
                                                                },
                                                                "descripcion": {
                                                                    "description": "Descripción detallada del programa",
                                                                    "type": "string",
                                                                    "example": "Fondos concursables para nuevos negocios"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "ano": "2023",
                                        "total_monto": 295608,
                                        "num": 8,
                                        "beneficios": [
                                            {
                                                "id_programa": 147,
                                                "monto": 40656,
                                                "fecha_recepcion": "09/11/2023",
                                                "fecha": "2023-11-09",
                                                "ano": "2023",
                                                "view": true,
                                                "ficha": {
                                                    "id": 922,
                                                    "nombre": "Emprende",
                                                    "id_programa": 147,
                                                    "url": "emprende",
                                                    "categoria": "trabajo",
                                                    "descripcion": "Fondos concursables para nuevos negocios"
                                                }
                                            }
                                        ]
                                    },
                                    {
                                        "ano": "2022",
                                        "total_monto": 150000,
                                        "num": 3,
                                        "beneficios": []
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al procesar beneficios: Error al obtener beneficios desde API externa"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "error": "Error al procesar beneficios: Error al obtener beneficios desde API externa"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Beneficios",
            "description": "Endpoints para gestión de beneficios procesados"
        }
    ]
}